# docs/swagger.yaml
definitions:
...
  db.User:
    properties:
      dateJoined:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      isStaff:
        type: boolean
      isSuperuser:
        type: boolean
      lastLogin:
        type: string
      lastName:
        type: string
      password:
        type: string
    type: object
...
  github_com_isaiaspereira307_gopvp_handlers_users_handler.ErrorResponse:
    properties:
      error_code:
        type: string
      message:
        type: string
    type: object
...
  handlers.DeleteUserResponse:
    properties:
      data:
        $ref: '#/definitions/db.User'
      message:
        type: string
    type: object
...
  handlers.ShowUserResponse:
    properties:
      data:
        $ref: '#/definitions/db.User'
      message:
        type: string
    type: object
...
  handlers.UpdateUserRequest:
    properties:
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      password:
        type: string
    required:
    - first_name
    - id
    - last_name
    - password
    type: object
  handlers.UpdateUserResponse:
    properties:
      data:
        $ref: '#/definitions/db.User'
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an user
      parameters:
      - description: Delete User Param
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DeleteUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_isaiaspereira307_gopvp_handlers_users_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_isaiaspereira307_gopvp_handlers_users_handler.ErrorResponse'
      summary: Delete an user
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Show an user
      parameters:
      - description: Show User Request
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ShowUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_isaiaspereira307_gopvp_handlers_users_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_isaiaspereira307_gopvp_handlers_users_handler.ErrorResponse'
      summary: Show user
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update an user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UpdateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_isaiaspereira307_gopvp_handlers_users_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_isaiaspereira307_gopvp_handlers_users_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_isaiaspereira307_gopvp_handlers_users_handler.ErrorResponse'
      summary: Update an user
      tags:
      - user
swagger: "2.0"